requirements.txt
accelerate==0.28.0
annotated-types==0.6.0
appdirs==1.4.4
bitsandbytes==0.43.0
clip @ git+https://github.com/openai/CLIP.git
chamfer-distance @ git+https://github.com/otaheri/chamfer_distance@dc9987dcf70888d387d96893ba1fb9ba9a333992
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
datasets==2.18.0
deepspeed==0.15.4
diffusers==0.20.0
dill==0.3.8
docker-pycreds==0.4.0
filelock==3.13.1
fsspec==2024.2.0
gitdb==4.0.11
gitpython==3.1.42
h5py==3.10.0
hjson==3.1.0
huggingface-hub==0.21.4
jinja2==3.1.3
markupsafe==2.1.5
matplotlib==3.9.2
meshio==5.3.5
mpmath==1.3.0
multiprocess==0.70.16
networkx==3.2.1
ninja==1.11.1.1
numpy==1.26.4
opencv-python==4.9.0.80
packaging==24.0
pandas==2.2.1
peft==0.9.0
pillow==10.2.0
plyfile==1.1
protobuf==4.25.3
psutil==5.9.8
py-cpuinfo==9.0.0
pyarrow==15.0.1
pyarrow-hotfix==0.6
pydantic==2.6.4
pydantic-core==2.16.3
pynvml==11.5.0
python-dateutil==2.9.0.post0
# conda install -c conda-forge pythonocc-core=7.7.2
pytz==2024.1
pyyaml==6.0.1
regex==2023.12.25
requests==2.31.0
safetensors==0.4.2
scipy==1.13.1
sentencepiece==0.2.0
sentry-sdk==1.42.0
setproctitle==1.3.3
smmap==5.0.1
sympy==1.12
tensorboard==2.16.2
tokenizers==0.15.2
torch==2.2.1
torchaudio==2.2.1
torchvision==0.17.1
tqdm==4.66.2
transformers==4.38.2
trimesh==4.6.0
tzdata==2024.1
urllib3==2.2.1
wandb==0.16.4
xxhash==3.4.1